requirements.txt
pygithub
lancedb
pyarrow
llama-cpp-python
huggingface_hub
sentencepiece
